requirements.txt
asgiref==3.8.1
Django==5.1.6
django-appconf==1.1.0
django-cleanup==9.0.0
django-environ==0.12.0
django-imagekit==5.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.1
djangorestframework==3.15.2
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
pilkit==3.0
pillow==11.1.0
pytz==2025.1
sqlparse==0.5.3
